pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unified-neural-pipeline"
version = "0.1.0"
description = "Target speaker diarization and multispeaker ASR pipeline."
readme = "README.md"
requires-python = ">=3.10"
keywords = ["speech", "diarization", "asr", "speaker-identification"]
authors = [
  { name = "Awetales Unified Neural Pipeline Team" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
  "Framework :: FastAPI",
  "Topic :: Multimedia :: Sound/Audio :: Analysis"
]
dependencies = [
  "torch>=2.2.0",
  "torchaudio>=2.2.0",
  "numpy>=1.24",
  "scipy>=1.11",
  "soundfile>=0.12",
  "librosa>=0.10",
  "numba>=0.58",
  "transformers>=4.40",
  "pyannote.audio>=3.0",
  "hydra-core>=1.3",
  "onnxruntime-gpu>=1.17",
  "onnx>=1.15",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "websockets>=12.0",
  "aiofiles>=23.2",
  "structlog>=24.1",
  "prometheus-client>=0.20",
  "orjson>=3.9",
  "srt>=3.5",
  "pandas>=2.2",
  "plotly>=5.20",
  "python-socketio>=5.11",
  "sounddevice>=0.4",
  "pydantic>=2.7",
  "pydantic-settings>=2.2"
]

[project.optional-dependencies]
dev = [
  "black>=24.4",
  "ruff>=0.4",
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "pytest-cov>=4.1",
  "mypy>=1.10",
  "pre-commit>=3.6",
  "rich>=13.7"
]

[project.urls]
Repository = "https://example.com/unified-neural-pipeline"
Documentation = "https://example.com/unified-neural-pipeline/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "S", "A", "C4", "PT", "SIM"]
ignore = ["S101", "S311"]
extend-exclude = ["data", "outputs"]

[tool.pytest.ini_options]
addopts = "--strict-markers --disable-warnings"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
namespace_packages = true


